+++ release_notes.md
@@ -1,5 +1,8 @@
 # Document Viewer release notes
 
+## NEXT
+* [TAB-33620] DCV | AU Modifications for Gift Card Load
+
 ## 1.54.0
 * [TAB-32272] DCV | AU document viewer