+++ release_notes.md
@@ -1,6 +1,7 @@
 # Document Viewer release notes
 
 ## NEXT
+* [TAB-33620] DCV | AU Modifications for Gift Card Load
 * [TAB-33679] DCV | AU | Add the word refund before the payment if the payment type is refund
 
 ## 1.54.1